requirements.txt
django
django-rest-framework
argon2-cffi
django-cors-headers